Build packages changes

+++ config/rpm/percona-toolkit.spec
@@ -41,6 +41,16 @@ find $RPM_BUILD_ROOT -type d -depth -exec rmdir {} 2>/dev/null ';'
 find $RPM_BUILD_ROOT -type f -name 'percona-toolkit.pod' -exec rm -f {} ';'
 chmod -R u+w $RPM_BUILD_ROOT/*
 
+%pre
+INSTALLED=`rpm -q percona-toolkit | awk -F'-' '{print $3}' | awk -F'.' '{print $1}'`
+if [ $INSTALLED < 3 ]; then
+    echo "** It is not possible to install percona-toolkit-3.0.0rc as it will replace already installed package"
+    echo "** Please note that percona-toolkit-3.0.0 is RC"
+    echo "** If you want to install it you need to remove already installed package"
+    exit 1
+fi 
+
+
 %clean
 rm -rf $RPM_BUILD_ROOT
